Replacing errant `ThrowingPlayerChoiceContext`s with appropriate passed parameters

+++ IntoTheSpireverseCode/Character/ShadowIronclad/Powers/GrapplePower.cs
@@ -21,7 +21,7 @@ public sealed class GrapplePower : ShadowPowerModel
         if (amount <= 0m || creature != Source) return;
         Flash();
         await CreatureCmdCompatibility.Damage(
-            new BlockingPlayerChoiceContext(), Owner, (decimal)Amount,
+            new BlockingPlayerChoiceContext(), Owner, Amount,
             ValueProp.Unpowered, Source, null, null);
     }
